added addLot function with corresponding methods

+++ app/Database/migrations/20221130085412_add_parking_lot_table.php
@@ -25,7 +25,7 @@ final class AddParkingLotTable extends AbstractMigration
         $table = $this->table('ParkingLots');
         $table->addColumn('address','string', ['limit' => 50])
             ->addColumn('space_count','int')
-            ->addColumn('hourly_rate', 'int')
+            ->addColumn('hourly_rate', 'float')
             ->create();
     }
 }
